Move SearchActions into actions folder
