+++ client/src/detail/DetailContainer.js
@@ -1,8 +1,8 @@
 import { connect } from 'react-redux'
 import { setFocus } from './DetailActions'
-import { fetchGranules, clearGranules } from '../search/SearchActions'
+import { fetchGranules, clearGranules } from '../actions/SearchActions'
 import { toggleSelection, clearSelections } from '../result/collections/CollectionsActions'
-import { clearSearch, triggerSearch, updateQuery } from '../search/SearchActions'
+import { clearSearch, triggerSearch, updateQuery } from '../actions/SearchActions'
 import { clearFacets } from '../search/facet/FacetActions'
 import { showCollections, showGranules } from '../actions/FlowActions'
 import Detail from './DetailComponent'
