+++ client/src/result/granules/list/GranuleListContainer.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux'
-import { toggleGranuleFocus } from '../../../search/SearchActions'
+import { toggleGranuleFocus } from '../../../actions/SearchActions'
 import GranuleList from './GranuleListComponent'
 
 const mapStateToProps = (state) => {
