+++ client/test/reducers/behavior/requestSpec.js
@@ -1,8 +1,8 @@
 import '../../specHelper'
 import Immutable from 'seamless-immutable'
 import { request, initialState } from '../../../src/reducers/behavior/request'
-import { fetchingGranules, fetchedGranules } from '../../../src/search/SearchActions'
-import { startSearch, completeSearch } from '../../../src/search/SearchActions'
+import { fetchingGranules, fetchedGranules } from '../../../src/actions/SearchActions'
+import { startSearch, completeSearch } from '../../../src/actions/SearchActions'
 
 describe('The request reducer', function () {
